+++ examples/wells_example.cpp
@@ -3,7 +3,7 @@
 #include <opm/core/utility/parameters/ParameterGroup.hpp>
 
 
-#include <opm/core/utility/initState.hpp>
+#include <opm/core/simulator/initState.hpp>
 #include <opm/core/simulator/SimulatorTimer.hpp>
 #include <opm/core/wells/WellsManager.hpp>
 #include <opm/core/grid/GridManager.hpp>
