+++ opm/core/transport/reorder/TransportSolverTwophaseReorder.cpp
@@ -21,7 +21,7 @@
 #include <opm/core/props/IncompPropertiesInterface.hpp>
 #include <opm/core/grid.h>
 #include <opm/core/transport/reorder/reordersequence.h>
-#include <opm/core/utility/ColumnExtract.hpp>
+#include <opm/core/grid/ColumnExtract.hpp>
 #include <opm/core/utility/RootFinders.hpp>
 #include <opm/core/utility/miscUtilities.hpp>
 #include <opm/core/pressure/tpfa/trans_tpfa.h>
